refactor: rename gstr3b data files
