+++ india_compliance/gst_india/report/hsn_wise_summary_of_inward_supplies/hsn_wise_summary_of_inward_supplies.py
@@ -3,7 +3,7 @@ from india_compliance.gst_india.report.hsn_wise_summary_of_outward_supplies.hsn_
     process_hsn_data,
     validate_filters,
 )
-from india_compliance.gst_india.utils.gstr3b.gstr3b_data import GSTR3BInvoices
+from india_compliance.gst_india.utils.gstr3b.gstr3b_inward_data import GSTR3BInvoices
 
 
 def execute(filters=None):
